backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-decouple
django-admin-rangefilter
drf-yasg
djangorestframework-simplejwt==5.5.0